change structures

+++ src/main/java/com/example/comicop/entity/Account.java
@@ -1,9 +1,7 @@
 package com.example.comicop.entity;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 import java.io.Serializable;
 
@@ -15,7 +13,7 @@ import java.io.Serializable;
 public class Account implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long userID;
+    private int userID;
     private String userName;
     private String password;
     private String email;
